% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_with_annotation.R
\name{load_data_with_annotation}
\alias{load_data_with_annotation}
\title{Load and align data with annotations}
\usage{
load_data_with_annotation(
  data,
  annotation,
  id_col = NULL,
  annotation_col = NULL
)
}
\arguments{
\item{data}{Either a data frame or a file path to a CSV/XLSX file containing subjects in rows and features in columns.}

\item{annotation}{Either a factor/vector, a data frame, or a file path to a CSV/XLSX.}

\item{id_col}{Optional name of column to use as rownames for both data and annotation. Default is NULL.}

\item{annotation_col}{Optional name of column in annotation data frame to use as the classification factor if there are multiple columns.}
}
\value{
A list with:
\describe{
\item{data}{A data frame of the input data}
\item{annotation}{A factor vector aligned to the data by rowname}
}
}
\description{
This function loads a data matrix and associated annotations from either files or data frames,
and returns them in a standardized format.
}
