% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_cv.R
\name{nested_cv}
\alias{nested_cv}
\title{Nested Cross Validation}
\usage{
nested_cv()
}
\description{
Performs a 5 fold nested cross validation on patientFeaturesData i.e. data_X and data_Y using 8 different machine learning models.
}
